Add files via upload
This is a zip file of an example out put.  It is coming out at around 30 meg at the moment and so I need to look into why it is coming out so big.  It should be around 10 meg tops.  I think it may be to do with the DPI settings I have.